sftp-s3-backup-tool/cdk/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
python-dotenv>=0.19.0
